n per loop
